Remove zip file after destroyed

+++ destroy.ps1
@@ -1,1 +1,5 @@
-terraform -chdir=terraform/ destroy -auto-approve
+if (Test-Path -Path terraform/blog-api.zip)
+{
+    terraform -chdir=terraform/ destroy -auto-approve
+    Remove-Item terraform/blog-api.zip
+}
